requirements.txt
pandas
pytest
streamlit
openai
numpy
scikit-learn
python-dotenv
